# Fix Benchmarks for go1.5