@paulmach 2015-10-10
Fixed the Visvalingam allocations
```
benchmark                           old ns/op     new ns/op     delta
BenchmarkVisvalingamThreshold-4     31662873      26382193      -16.68%
BenchmarkVisvalingamKeep-4          15825549      11250108      -28.91%

benchmark                           old allocs     new allocs     delta
BenchmarkVisvalingamThreshold-4     73755          8              -99.99%
BenchmarkVisvalingamKeep-4          73757          17             -99.98%

benchmark                           old bytes     new bytes     delta
BenchmarkVisvalingamThreshold-4     4152311       3569745       -14.03%
BenchmarkVisvalingamKeep-4          4875584       4293495       -11.94%
```